Add API to open CL doc
